requirements.txt
addict==2.4.0
aiofiles==24.1.0
albumentations==1.4.2
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
coverage==7.9.1
crcmod==1.7
cryptography==45.0.4
cycler==0.12.1
einops==0.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
ffmpy==0.6.0
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2024.6.1
gdown==4.4.0
gradio==5.34.1
gradio_client==1.10.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iniconfig==2.1.0
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
-e git+https://github.com/open-mmlab/mmcv.git@90d83c94cfb967ef162c449faf559616f31f28c2#egg=mmcv
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
onnx==1.18.0
onnxoptimizer==0.3.13
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pillow==10.2.0
platformdirs==4.3.8
pluggy==1.6.0
prefetch_generator==1.0.1
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.0
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyTurboJPEG==1.8.0
pytz==2023.4
PyYAML==6.0.2
requests==2.28.2
rich==13.4.2
ruff==0.12.0
safehttpx==0.1.6
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
semantic-version==2.10.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
sympy==1.13.3
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.65.2
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
vietocr==0.3.13
websockets==15.0.1
yapf==0.43.0
